+++ src/Object/Content.php
@@ -256,7 +256,7 @@ class Content implements
         return $this->localeField;
     }
 
-    public function setLocaleField(string $localeField): TranslatableInterface
+    public function setLocaleField(?string $localeField): TranslatableInterface
     {
         $this->localeField = $localeField;