+++ tests/universal/Object/ItemTest.php
@@ -245,6 +245,19 @@ class ItemTest extends TestCase
         );
     }
 
+    public function testSetLocaleFieldToNull()
+    {
+        $Object = $this->buildObject();
+        self::assertInstanceOf(
+            \get_class($Object),
+            $Object->setLocaleField(null)
+        );
+
+        self::assertNull(
+            $Object->getLocaleField()
+        );
+    }
+
     public function testSetLocaleFieldExceptionOnBadArgument()
     {
         $this->expectException(\Throwable::class);